[+] Move refresh playlist to MusicPlayer
